advisories: add patched package with bottlerocket- prefix name
In any Bottlerocket variant built with the combination of twoliter version
< 0.6.0 and the bottlerocket-kernel-kit (any version), application
inventory for the variant will not strip the bottlerocket- prefix from
the name of packages sourced from the kernel-kit.

Remediate this by adding bottlerocket- prefixed package names to
advisories applicable to kernel-kit < 1.2.0. kernel-kit v1.2.0 built
with twoliter v0.7.3, which included a kit metadata version bump to
signal a breaking change.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>