kernel-6.12: adjust zram config to prefer LZ4
LZ4 is faster at compression and decompression than LZO, and is also
used by erofs for compressing files on the root filesystem.

Using the same compression format makes it easier to reason about the
trade-off between evicting clean pages and swapping anonymous pages:
both incur the penalty of LZ4 decompression to access the memory
again, but the clean pages must also be read back from the underlying
block device.

We also use LZ4 in other contexts, such as compressing the artifacts
for in-place updates, while LZO is not used anywhere else.

Signed-off-by: Ben Cressey <bcressey@amazon.com>