kernel-6.12: unset option for kheaders.tar.xz
Prior to the availability of BTF, making kernel headers available via
/sys/kernel/kheaders.tar.xz allowed BPF programs to be built for the
running kernel. For modern tools, this is superseded by the BTF data
available in /sys/kernel/btf/vmlinux, and is no longer required.

The xz-compressed headers are a 2 MiB blob that cannot be compressed
any further, which increases the kernel size and slows boot because
the additional data must all be read and verified.

Signed-off-by: Ben Cressey <bcressey@amazon.com>