kernel-6.12: simplify devel packaging
Do not make the kernel development files available at runtime, since
any kernel modules built this way are not compatible with a kernel in
lockdown mode. Continuing to support this interface prevents us from
enabling other features such as LoadPin that would further restrict
the ability to load kernel modules from unexpected sources.

There are better ways to build out-of-tree kernel modules now, using
Twoliter and custom packages that depend on the published kernel kits
that correspond to the target downstream images.

Remove the various archives for kernel development files, and replace
all of them with a single subpackage that includes the files in a way
that's easy for downstream packages to consume as a build dependency.

Signed-off-by: Ben Cressey <bcressey@amazon.com>