kernel-6.1, -6.12: disable PCR 9 measurement
On aarch64, the kernel measures the command line from the bootloader
into PCR 9 during early boot. As the sequence below shows, this takes
place before bootconfig can be parsed.

  efi_pe_entry()
    ↓
  efi_handle_cmdline(image, &cmdline_ptr)
    ↓
  efi_convert_cmdline(image, &cmdline_size)
    - Measures boot_command_line into PCR 9
    ↓
  handle_kernel_image()
    ↓
  efi_stub_common()
    ↓
  efi_boot_kernel()
    ↓
  efi_enter_kernel()
    ↓
  [architecture-specific early setup]
    ↓
  start_kernel()
    ↓
  setup_arch(&command_line)
    ↓
  setup_boot_config()
    - Extracts bootconfig from initrd
    - Sets up extra_command_line
    ↓
  setup_command_line(command_line)
    - Concatenates extra_command_line + boot_command_line

Since an incomplete measurement results in a PCR value that's hard to
predict, disable the kernel's measurement and rely on the verified
userspace to measure the final value from /proc/cmdline.

Signed-off-by: Ben Cressey <bcressey@amazon.com>