kernel-6.12: Add additional Neuron module
Add a second neuron module so that we can choose the appropriate driver
based upon the detected hardware. This provides both neuron_2_21 and
neuron_latest in the data directory which is not discoverable by depmod.
The mechanism for loading the driver is driverdog link and load modules
for neuron.

There is now two scripts for checking the neuron driver, one for 2.21.*
and one for latest.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>