kernel-6.1: include efa from efa-installer
The EFA driver in the Amazon Linux kernel can have some delay between
when the installer is released and it making it into the kernel tree.
This adds a build step to update the driver from the installer to get
the latest changes as soon as the installer is available.

As part of this change, the `efa.cmake` file was updated to do a serial
execution of the tests to validate the available configurations in the
kernel sources. This is a problem unique to Bottlerocket due to how out
of tree kmod builds are treated.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>