kernel-6.12: test EFA configurations in parallel
In e0ab5af, a patch was introduced to execute EFA's configuration
validations in serial. The problem the patch fixed was self-inflicted by
the patch 1001-Makefile-add-prepare-target-for-external-modules.patch,
which is required in Bottlerocket to support building out of tree kernel
modules.

Update said patch to allow the parallel EFA configuration validations,
and drop the patch to make the validations serial.

Signed-off-by: Arnaldo Garcia Rincon <agarrcia@amazon.com>